Bug Fixing

+++ gameengine/src/main/java/de/mirkosertic/gameengine/arcaderacer/Road.java
@@ -4,13 +4,13 @@ import de.mirkosertic.gameengine.type.Color;
 
 public class Road {
 
-    public final float positionLeft;
-    public final float positionRight;
+    public final double positionLeft;
+    public final double positionRight;
     public final int curveFactor;
     public final Color roadColor;
     public final Curb[] curbs;
 
-    public Road(float aPositionLeft, float aPositionRight, int aCurveFactor, Color aRoadColor, Curb[] aCurbs) {
+    public Road(double aPositionLeft, double aPositionRight, int aCurveFactor, Color aRoadColor, Curb[] aCurbs) {
         positionLeft = aPositionLeft;
         positionRight = aPositionRight;
         curbs = aCurbs;
